+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -67,6 +67,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added controls for the users to manually feed the ray tracing acceleration structure that should be used for a given camera (case 1370678).
 - Depth of Field is now disabled in orthographic cameras - it was using the hidden perspective settings (case 1372582).
 - Modified HDRP to use common FSR logic from SRP core.
+- Exposed the volumetric clouds transmittance for volumetric clouds (case 1373283).
 
 ## [13.1.0] - 2021-09-24